+++ cvmfs/CMakeLists.txt
@@ -36,6 +36,7 @@ set (CVMFS_CLIENT_SOURCES
   cache.cc cache.h
   cache_posix.cc cache_posix.h
   cache_ram.cc cache_ram.h
+  cache_tiered.cc cache_tiered.h
   catalog.cc catalog.h
   catalog_counters.cc catalog_counters.h catalog_counters_impl.h
   catalog_mgr.h catalog_mgr_impl.h
@@ -85,7 +86,6 @@ set (CVMFS_CLIENT_SOURCES
   sqlitemem.cc sqlitemem.h
   sqlitevfs.cc sqlitevfs.h
   statistics.cc statistics.h
-  tiered_cache.cc tiered_cache.h
   tracer.cc tracer.h
   uid_map.h
   uuid.cc uuid.h
