+++ cvmfs/mountpoint.cc
@@ -35,6 +35,7 @@
 #include "cache.h"
 #include "cache_posix.h"
 #include "cache_ram.h"
+#include "cache_tiered.h"
 #include "catalog.h"
 #include "catalog_mgr_client.h"
 #include "clientctx.h"
@@ -60,7 +61,6 @@
 #include "sqlitemem.h"
 #include "sqlitevfs.h"
 #include "statistics.h"
-#include "tiered_cache.h"
 #include "tracer.h"
 #include "util/pointer.h"
 #include "util/posix.h"
